backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
firebase-admin==6.5.0
web3==6.11.3
eth-account==0.10.0
cryptography==41.0.7
pydub==0.25.1
jiwer==3.0.3
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
requests==2.31.0
Levenshtein==0.23.0
django-filter==23.5
drf-spectacular==0.26.5
